Create twoSum
Two Sum uses a hashmap  to map the no. in the array with its respective index. we then check the difference between the target and numer is there in the array or not. if it is there that means there is the no. and we return a new created array with the diff(a no. ) and i (another no.).